fix: implement nonce-based CSP and suppress ZAP-10031 false positive
- Generate cryptographic nonce per-request in middleware for CSP script-src
- Remove unsafe-inline from script-src, replace with nonce-based validation
- Move CSP construction from static next.config.ts to dynamic middleware
- Read x-nonce in layout for Next.js to apply to generated scripts
- Update .zap/rules.tsv: add 10031 (false positive), update 10055 comment

This eliminates ZAP-10055 (script-src unsafe-inline) real security finding
and adds 10031 (false positive URL param in HTML attr) to suppressions.
